Cleanup refactor of styles code.
